+++ src/common/domain/exceptions/invalid-id.exception.ts
