+++ src/common/domain/models/aggregate-root.spec.ts
@@ -1,7 +1,7 @@
 import * as faker from 'faker';
-import { AggregateRoot } from './AggregateRoot';
-import { Entity } from './Entity';
-import { Id } from './Id';
+import { AggregateRoot } from './aggregate-root';
+import { Entity } from './entity';
+import { Id } from './id';
 
 describe('Aggregate root', () => {
   it('is an entity', () => {
