+++ src/common/domain/models/aggregate-root.ts
@@ -1,5 +1,5 @@
-import { Entity } from './Entity';
-import { Id } from './Id';
+import { Entity } from './entity';
+import { Id } from './id';
 
 export abstract class AggregateRoot extends Entity {
   protected constructor(readonly id: Id) {
