+++ src/common/domain/models/index.ts
@@ -1,4 +1,4 @@
-export * from './AggregateRoot';
-export * from './Entity';
-export * from './Id';
-export * from './ValueObject';
+export * from './aggregate-root';
+export * from './entity';
+export * from './id';
+export * from './value-object';
