+++ src/common/domain/models/value-object.ts
